Stop course enrollment emails & redirect to course page after enroll
