+++ openedx/core/djangoapps/timed_notification/notifications.py
@@ -14,7 +14,7 @@ def enrollment_confirmation(sender, event=None, user=None, **kwargs):
             'course_name': course.display_name,
             'course_link': get_course_link(course_id=course.id)
         }
-        send_course_notification_email(course=course,
-                                       mako_template_path='timed_notification/enrollment_confirmation.txt',
-                                       context=context,
-                                       to_list=[user])
+        # send_course_notification_email(course=course,
+        #                                mako_template_path='timed_notification/enrollment_confirmation.txt',
+        #                                context=context,
+        #                                to_list=[user])
